+++ packages/eslint-config/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./config/base.js";
 export * from "./config/i18next.js";
+export * from "./config/react.js";
 export * from "./config/tailwind.js";
 export * from "./config/tanstack-router.js";